Implement Computer move into the game
now when player1 click a cell, computer will respond with either
attack movement or defend movement, so this computer should be good
enough to at least defend the game.

Because the nature of tic tac toe, second playing game will be in
disadvantage, so as long as player not make a mistake they should be
able to win from the computer.